fix ConcurrentModificationError in AttributedSpans.hasAttributionsWithin
AttributedSpans.hasAttributionsWithin iterates over a Set and removes entries while iterating on it. This throws an ConcurrentModificationError and therefore does not work. I guess the Set iterator does not know it's iteration order if you remove things from under it's feet. 

I think it may not have been noticed until now as most people only check for one Attribution and therefore exit the function before the error can be noticed.